+++ CHANGELOG.md
@@ -18,6 +18,14 @@ For older changes see the [archived Singularity change log](https://github.com/a
   directly from Dockerfile, without having to use Docker/Podman but only BuildKit.
   The full buildkit log file is included in the image, for traceability.
   It is also showed on the console, as a progress update while building.
+- Support hosts that have `/etc/resolv.conf` pointing to a symlink under
+  `/run`, such as those hosts that are running `systemd-resolved`.  In
+  this case, the symlink is copied into the container and the parent
+  directory of the target of the symlink is bind-mounted from the host.
+  The result is that even if the target of the symlink is replaced with
+  a new file, the container sees the update in `/etc/resolv.conf`.
+- Add `/etc/resolv.conf` to the list of host paths that can be prevented
+  from automatic import into the container with the `--no-mount` option.
 - Add support for selective mounting of Intel(R) Gaudi accelerators.
   This feature is only for use in combination with a minimal `/dev` directory,
   selected either with the `--contain` flag or by configuring `mount dev`